Deletion of material now separates cases where material is used and not used

+++ pages/api/deleteMaterial.php
@@ -16,11 +16,30 @@ if(!isset($_GET["id"])){
 }
 
 $email = getSessionEmail();
-$query = "DELETE FROM Materiale WHERE id=? AND idVenditore=?";
-#execute
-$stmt = mysqli_prepare($connection, $query);
-mysqli_stmt_bind_param($stmt,"is", $_GET["id"],$email);
+//controlla se il prodotto ha una variante
+$check_query = "SELECT COUNT(idProdotto) AS total FROM Variante WHERE idMateriale=?";
+$stmt = mysqli_prepare($connection, $check_query);
+mysqli_stmt_bind_param($stmt,"i", $_GET["id"]);
 mysqli_stmt_execute($stmt);
+$result = mysqli_stmt_get_result($stmt);
+$count = mysqli_fetch_assoc($result);
+
+
+if($count["total"]==0){
+    $query = "DELETE FROM Materiale WHERE id=? AND idVenditore=?";
+    #execute
+    $stmt = mysqli_prepare($connection, $query);
+    mysqli_stmt_bind_param($stmt,"is", $_GET["id"],$email);
+    mysqli_stmt_execute($stmt);
+}
+else{
+    $query = "UPDATE Materiale SET visibile = 0 WHERE id=?";
+    #execute
+    $stmt = mysqli_prepare($connection, $query);
+    mysqli_stmt_bind_param($stmt,"i", $_GET["id"]);
+    mysqli_stmt_execute($stmt);
+}
+
 header("Location: /sellerHome.php");
 
 ?>
